2.3.0 积分商城 购物车和提交订单金额的bug

+++ app/src/main/java/com/ksfc/newfarmer/protocol/beans/AddGiftOrderResult.java
@@ -72,7 +72,7 @@ public class AddGiftOrderResult extends ResponseResult {
             public String category;
             public String thumbnail;
             public int points;
-            public int marketPrice;
+            public float marketPrice;
             public int online;
             public int soldout;
         }
